chore: Add NameSpace.swift for move constants enum View Controller에서 작성한 StoryboardID StoryboardName AlertTitle 생성한 NameSpace파일로 이동
